Introduce common index with retries and move group handler
Indexing events should not be executed immediately, but rather wait for
the relevant git data to be replicated first. Because of that they all
share the same structure, they all need a way of checking whether data
is up-to-date, they all need a scheduler to retry, they all need to read
the same [index] configuration.

Create a base class for index handler and move group index handler to
use it. Subsequent changes will address other index handlers.

Change-Id: Ic46bef39bc72004dc2f5a39eac882db8aa414c55