Fix cache eviction for projects cache
When the project name contains a slash, the cache eviction router
throws a MalformedJsonException.

For projects cache pass the key without JSON transcoding, preventing
the problem of breakage when the string value is not a valid JSON
element.

Also rename the CacheKeyJsonParser.fromJson method to
CacheKeyJsonParser.from to reflect the fact that cache key value is not
always a JSON element.

Bug: Issue 14564
Change-Id: I6ce9471bd96456c0950e321d0c694ec2f9ac254c