Introduce local (docker) setup for e2e tests
This is the initial commit that creates local (docker based) setup
to run e2e tests. For now it just starts the setup with:

  e2e-tests/test.sh

It provides only the very basic test: start setup and check if all
required plugins were loaded.

See `e2e-tests/README.md` for requirements and details about the
test setup and call:

  e2e-tests/test.sh --help

for all configuration options.

Bug: Issue 14509
Change-Id: Ibb123a5e100d8abedb9035ba02ecb95acfab5ef8