Leverage stream events publishing from the events-broker
Events-broker now provides a class to handle the publishing of gerrit
stream events in a centralized way.

In order to maintain back-compatibility with the fact that events-broker
implementations were publishing stream events to a `gerrit` topic, an
new optional `broker.gerritEventsTopic` has been introduced.

The need for the additional `gerrit` topic can be removed once [1] has
been completed.

[1] https://bugs.chromium.org/p/gerrit/issues/detail?id=14835

Bug: Issue 14834
Change-Id: Iafe5a8155a98746191926286d3618beabf3aa452