Deprecate STREAM_EVENT_TOPIC
STREAM_EVENT_TOPIC was originally introduced to allow separating the
stream of messages of type EventMessage, so that they could be stamped
by an instance-id and thus underatnd where they originated.

Change I577ceb86 finally removed the need for this distinction, so that
now only the `GERRIT_TOPIC` event should be used.

Annotate the STREAM_EVENT_TOPIC as deprecated to signal that only the
`GERRIT_TOPIC` should be used instead.

Change-Id: I4913da1bdc58268103f02825f9e4cebae76c23c7