Add local projects filtering to the message receivers
On stable-3.5 global/local projects filtering is moved from producer to
receiver side. This means that messages are sent for both local and
global projects. To allow rolling upgrades add global/local projects
filtering on receiving side. This ensure that messages produced for
local projects on stable-3.5 node are filtered out on stable-3.4 as well.

Bug: Issue 14907
Change-Id: I95dc0e71c823ff18a962c1d11daedb75c45e01bc