Extract replication status logic from subscriber metrics
Extract the logic to update and retrieve the replication status of
projects, so that the same logic can be used elsewhere.

This is useful as a preparation for the introduction of an HTTP endpoint
to retrieve project replication status information on demand.

This will be done as a follow up change.

Bug: Issue 15309
Change-Id: I9b82ef46e431420f5c3e80a73be89da0b8b0b024