Introduce persistent cache for replication-status
ReplicationStatus keeps track of the replication lags associated to
projects in a multi-site set-up, so that it is possible to understand
how much behind a particular project is falling due to replication.

This information however is lost upon gerrit restart.

Persist replication status information by using a persistent cache to
populate in-memory replication-status at startup.

Bug: Issue 15310
Change-Id: I86c40684439ec5a457866f56f4cb9157c4af0092