Delete replication metrics upon project deletion
Change I7259f520 introduced the handling of
ProjectDeletionReplicationSucceededEvent events to remove the metrics of
deleted projects upon successful replication of project deletions.

However the metrics were not removed from the gerrit instance that
initiated the replication, since it did not consume the
ProjectDeletionReplicationSucceededEvent produced by itself.

Register ReplicationStatus as a ProjectDeletedListener so that local
replication metrics can be removed as soon as the project is deleted
from the local repository.

Bug: Issue 12958
Change-Id: I5c3f2e51fddace3d76cf5b94ecd19dc4c3004440