Fix exception when consuming index events of deleted changes
When producing ChangeIndex events, the producing side does not set the
name of the project in the event[1]. This is due to the `onChangeDeleted`
interface, which can only provide the change Id and not the project name
(see here [2]).

This, in turn, causes the global-refdb library to throw an exception,
when attempting to filter projects that the receiver is interested in
consuming (see here[3]), since it cannot tell which project the event is
associated to.

The `ChangeIndexEvent` cannot be hydrated with the name of the project at
production time, because the change has been deleted already, so
attempts to lookup the project name would fail.

The receiving side, on the other hand, can still find the deleted
change in the index and thus recover the name of the project before
asking the global-refdb to match the project name.

[1] https://gerrit.googlesource.com/plugins/multi-site/+/refs/heads/master/src/main/java/com/googlesource/gerrit/plugins/multisite/index/IndexEventHandler.java#136

[2]https://gerrit.googlesource.com/gerrit/+/refs/heads/master/java/com/google/gerrit/extensions/events/ChangeIndexedListener.java#46

[3]https://review.gerrithub.io/plugins/gitiles/GerritForge/global-refdb/+/refs/heads/master/src/main/java/com/gerritforge/gerrit/globalrefdb/validation/ProjectsFilter.java#126

Bug: Issue 15433
Change-Id: I1914eb5b73e5cca5eedfbb47c433b2fc27583329