Add kafka-rest proxy to the docker-compose env
Show how to use the kafka-rest proxy in front
of a kafka broker.

This can be useful when setting up real production
environment where the kafka broker isn't accessible
externally or when the brokers' DNS names cannot be
exposed externally.

Use the kafka-rest proxy from the Gerrit intsances
so that the E2E test can verify a real-life use-case.

Change-Id: If1503e1ffbbb22221b9e7d1171115545fc7256de