Add kafka-rest proxy to the docker-compose env
Show how to use the kafka-rest proxy in front
of a kafka broker.

This can be useful when setting up real production
environment where the kafka broker isn't accessible
externally (e.g. AWS's MSK [1]) or when the brokers' DNS
names cannot be exposed externally.

Use the kafka-rest proxy provided by Confluent [2] from the Gerrit instances
so that the E2E test can verify a real-life use-case.

The use of REST-Proxies is one of the solution of how to connect
to AWS's MSK [3], it makes sense to have it as one of the standard
ways the multi-site setup is validated.

[1] https://aws.amazon.com/msk/
[2] https://docs.confluent.io/platform/current/kafka-rest/index.html 
[3] https://docs.aws.amazon.com/msk/latest/developerguide/aws-access.html

Change-Id: If1503e1ffbbb22221b9e7d1171115545fc7256de