Inject ProjectVersionRefUpdate as Provider<>
The ProjectVersionRefUpdate relies on the injection of
the GerritInstanceId which could be delayed in the bootstrap
of Gerrit injectors chain at startup.

Starting from v3.5, the GerritInstanceId, an optional injection,
could still be null when the ReplicationStatus instance is created
causing a Guice instantiation failure.

Using a Provider<> resolves the issue and allows the resolution
of the GerritInstanceId at runtime when all the Guice modules
are instantiated.

This change is applied in stable-3.3 for minimizing the conflicts
of the merges up to master, as it is compatible with all branches
even if not necessarily needed on stable-3.{3,4}.

Change-Id: I638988205a004733e6531e6ac4126e2ba652ac06