Exclude event-broker and global-refdb from shaded plugin jar
multi-site plugin has only compile time dependency on event-broker and
global-refdb modules. They are provided in runtime in <gerrit_site>/lib
directory. That's why it should be safe to exclude them from the shaded
jar.

This is also done similarly to the replication plugin, that is only
compile time dependency.

Change-Id: I682516ff764ae9b7f24a5784adedae7af54ab647