Reload latest SHA1 from local when filtering RemoteRefUpdate
The fix I579f277635 to the Issue 12425 (racy conditions when pushing refs)
has created a situation where refs are pushed even when they
are potentially out of sync.

The problem happened when a local ref was aligned with the global-refdb
but the SHA1 contained in the ref-update event was not, and therefore the
generated RemoteRefUpdate caused a remote ref to be winded back to
one commit.

If the local SHA1 of a ref has changed and is aligned with the global-refdb
then it can be used for push, and the RemoteRefUpdate needs to be updated
with the newer SHA1 for preventing the rewind of the branch to an earlier
point.

This completes the fix originally introduced for sorting out missed replication
events due to racy conditions when checking if the RemoteRefUpdate is
up-to-date with the global-refdb.

Issue: 12425
Change-Id: Ic9bd23694640bd8bfbf875876d1559713cf6c262