Allow auto-refresh of replication lags for outdated repos
Enable the configuration of the replication lag refresh period
for refreshing the entries associated with projects that lags behind.

This allows to fix caching issues when the local filesystem has been
replicated but Gerrit could have failed to refresh the associated
metrics because of filesystem caching issues (e.g. NFS).

Change-Id: I9563be386b2245bb88f08812802d0bd914c5fd9a