Implement pull-replication fetch filter for avoiding split-brain
Define a concrete implementation of the ReplicationFetchFilter which
allows to prevent split-brain by filtering out all refs that are
already in-sync with the global-refdb.

This is dependent from I19d586381 which introduces the optional
interface ReplicationFetchFilter, implemented and bound in the
multi-site's validation module.

Allow to use the multi-site as libModule for leveraging the
pull-replication fetch filter only and having everything else
disabled as the config below:

[cache]
  synchronize = false
[event]
  synchronize = false
[index]
  synchronize = false

Also adapt the Jenkinsfile to allow the checkout of the pull-replication
plugin as dependency.

Depends-On: https://gerrit-review.googlesource.com/c/gerrit-ci-scripts/+/366460
Bug: Issue 15637
Change-Id: If60f5f5eaf700076e954004a4037cc70a8def2f1