Implement pull-replication fetch filter for avoiding split-brain
Define a concrete implementation of the ReplicationFetchFilter which
allows to prevent split-brain by filtering out all refs that are
already in-sync with the global-refdb.

This is dependent from I19d586381 which introduces the optional
interface ReplicationFetchFilter, implemented and bound in the
multi-site's validation module.

Also adapt the Jenkinsfile to allow the checkout of the pull-replication
plugin as dependency and bump the global-refdb dependency to 3.4.8.2.

Depends-On: https://gerrit-review.googlesource.com/c/gerrit-ci-scripts/+/366460
Bug: Issue 15637
Change-Id: If60f5f5eaf700076e954004a4037cc70a8def2f1