Allow the installation of push/fetch replication filters
Define explicit configuration settings in multisite.config
for installing the push/fetch replication filter modules.

The installation as Gerrit libModule would not work because of
the need to use a child injector that has all the proper fields
and optional injections resolved at the plugin startup phase.

Adapt the multi-site plugin documentation to reflect the new
settings.

Change-Id: Ic90adbe6dda515965aa684c0ba638a2da68b373c