Filter out non existing refs from fetch
Exclude ref from fetch task when ref is not present in the local
repository or has been removed in the global-refdb and therefore is
set to zeros.

P.S. The situation where the entry is not on the global-refdb is
in doubt, because we don't really know if fetching the refs would
succeed or not. The best approach is to fetch anyway and leave the
Git protocol to decide. That allows the fetching from instances when
the global-refdb has been initialised or for new repos that have
been imported but are not yet fully tracked on the global-refdb.
Furthermore, the global-refdb tracks only the mutable ref so
we cannot assume that everything is tracked.

Bug: Issue 16824
Change-Id: I06a3bb73a805046ec5a8bc582586e7db4d213171