Do not forward indexing events from pull-replication/apply-object
The pull-replication plugin supports synchronous replication, where
the Git blobs and refs are sent over the wire, applied at rest and,
if it is a /meta ref, the corresponding change is automatically reindexed.

Because the reindex at rest is a consequence of an remote /meta ref
update, the change is already indexed at source and therefore there is
no need to issue additional reindexing events.

Bug: Issue 16934
Depends-on: Change 373739
Change-Id: I0a28e302eeb417ed99fd500604b8be83ef3c3556