Manage FetchRefReplicatedEventHandler inside multi-site
The FetchRefReplicatedEventHandler logic is moved from pull-replication
into multi-site, so that the local reindexing of changes following a
pull-replication ref-replicated stream event would cause a loop of
indexing forwarding events.

This move also make the overall architecture more consistent, where
the responsibility of propagating the change reindexing is kept inside
multi-site.

This is the proper solution to the Issue 16934 which was temporary
workarounded in I0a28e302e.

Bug: Issue 16934
Change-Id: I6f9c8627b5b07acbf82f85f702240592b8ee263a