Move EventHandler binding from lib module to plugin module
Lib module binding happens before the plugins bindings. This means
that all EventsListeners defined in lib modules are executed before
EventsListeners defined in plugin module.

When multi-site works with pull-replication or other plugins,
the EventHandler listener should be executed respecting the priority
of loading and execution of the plugins, defined by the plugin filename
alphabetical ordering.

Historically, for sake of simplicity, the EventModule was put inside the
libModule of the multi-site, even if it was originally as part of the plugin
module. There is no special reason for the EventModule to be loaded
early on as libModule, and there are even drawbacks on the plugin
priority execution.

Move back the EventModule where it was originally.

Change-Id: I02355a50c896e7b971686873f46d8e3917ce6630