Use msec-level precision for refs/multi-site/version value
Very active repositories may have multiple updates within the same
second which deserve to be traced in the project version. Previously
with the maximum precision of 1 second there was the risk to lose
track of missed replication updates for multiple modifications.

The warning 'NOT Updating project <> in shared ref-db' was noticed
many times, which it was a cause of concern for the lack of
traceability of potential issues.

NOTE: This has a consequence on the replication-lag metric, which
has now a more accurate version which tracks at msec level.

TODO: Once this change is merged up to master, the legacy metric
'multi_site/subscriber/subscriber_replication_status/sec_behind'
must be removed because it would be confusing.

Change-Id: I9cc78ae0ce67b90ec1a82091b3ca60294086350e