Fix pull replication fetch filter for non-existing refs on global-refdb
The refs that are not created yet on the global-refdb were supposed to
be always fetch-able by pull-replication, because considered as new.

However, an invalid mocking approach to testing has lead the underlying
bug to be left unnoticed in the code base. Fix the mocking approach
and also the overall behaviour of pull-replication filtering in absence
of a global-refdb record.

Change-Id: I11df9b79b48d31e9a36a86e35266403ef6a3d940