Forward project list update event to the other master
In a multi-site setup with HA, when a new project is created in one
site, a project-list-update event is published to the broker so that
the cache is updated when the project is replicated. The node on the
other site that consumes that message will update their project list
cache, and they will not forward it to the other master. This means
that the project-list cache of the other master will not be updated,
and therefore the project will not show up in that master's UI.

Fire a project-created event on the node that received the
project-list-update msg, right after the cache is updated. This will
make the event available to the `NewProjectCreatedListener` listeners
which are registered. The multi-site `ProjectListUpdateHandler` is
such a listener, this will ignore the message as it considers it a
forwarded event (because the context is set when the event is fired).
The HA `ProjectListUpdateHandler` however will process the event and
forward it to the other master, who will update their cache.

Bug: Issue 40015518
Release-Notes: Forward project list update event to the other master
Change-Id: Id0086835adc6c9eb3231c62dff085b216195ba25