Use put operation for project version update
Project version update is used to generate metrics
about the project replication state. This information
doesn't have to be milliseconds precise, if two updates
happened within a short period of time we do not have to
guarantee that the latest one will be stored as long as
the project version ref is updated to one of those two.
This means that using `compareAndPut` operation is not
necessary and more lightweight `put` operation can be
used instead.

Bug: Issue 297440086

Change-Id: I6d3413f86c55f5fc9ab6323d30628e58ad74a137