Avoid concurrency/locking on refs/multi-site/version
Large mono-repos with huge ref-updates concurrency can
lead to excessive locking and updates on the
refs/multi-site/version.

Use a compare-and-set in-memory operations for avoiding
updating the project version when concurrent threads
are doing the same operation.

If two threads are attempting to update refs/multi-site/version
at the same time, accepting only one of the two updates
is good enough because it would have most likely be
same or similar version to the failing thread.

Also use the filesystem racy resolution time for avoiding to
issue multiple updates on the same ref within the racy
period that would always result in a filesystem LOCK
failure and therefore won't be stored anyway.

If the repository doesn't have a racy period (e.g. in-memory)
then assume 1s as the default resolution.

Change-Id: I7f12dede1c68bfec557a6b08b2470d2eb5e91e68