Disable the thread-local refs caching
The thread-local refs caching opens the risk to have a split
brain situation within the same JVM, because of two threads
not being able to see the refs modifications made elsewhere
in the Gerrit.

The thread-local caching of the refDabase was introduced in
Gerrit v3.8.0 with [1] and enabled by default. Although the caching
improves read performance, it would also cause data inconsistency
between threads.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/360646

Bug: Issue 309098227
Change-Id: I967cc56f0cc080d832d6623700f30efdf6468380