Delegate publishing of events to events-broker
Swap multi-site's broker api wrapper with the new wrapper from
`events-broker`, which was introduced in [1].

Upgrading the metrics will be done in a separate change, as such for now
the responsibility of publishing metrics is delegated to the
`BrokerForwarder`, who triggers the publishing of the messages.

Finally, the logic that decides if an event should be skipped from
publishing also moves into the `BrokerForwarder`, as it's multi-site
specific.

[1] https://gerrit-review.googlesource.com/c/modules/events-broker/+/393274

Bug: Issue 294904654
Change-Id: If837155e2a878bf07c018a943e9fe0fb18d454c4