Ignore refs/multi-site/version on global-refdb in replication filter
When I509dd7075b0 stopped updating the global-redb in this stable-3.4
branch, it did not take into consideration that existing setups
may have plenty of refs/multi-site/version already stored on them.

Stopping the update of those refs on the global-refdb had a negative
impact on the consistency of the replication of refs/multi-site/version
across the multi-site clusters, causing false alarms of replication lags.

Consider the refs/multi-site/version on global-refdb as non-existent
and skip altogether the associated checks if the associated local refs.
This change makes effectively like the global-refdb keys were effectively
removed and solve the temporary issue.

Change-Id: I8b938675f54d18542e34b735ff1e6d63f2ab2a31