Update patch set 2
Abandoned

For more context and details, see the recording in [1]. Below is a brief summary:

This change won't be merged, because there is no value in moving the `BrokerApiWrapper` to events-broker. The reason being that updating the message log file will still be happening in 2 different places: One for stream events in StreamEventsPublisher, and one for all other events in BrokerApiWrapper.

StreamEventsPublisher currently delegates publishing to an implementation of BrokerApi (be it one in events-kafka, events-gcloud-pubsub etc) via a DynamicItem<BrokerApi>, and BrokerApiWrapper is an implementation of BrokerApiWrapper. So in theory we could (in StreamEventsPublisher) swap the DynamicItem<BrokerApi> with a BrokerApiWrapper instance. However, this will introduce problems, as both classes require an Executor which serves a different purpose; for stream events, it needs to be single threaded and await for the publishing to finish as ordering is very important (see [2]), but for other types of events this is not necessarily the case.

Therefore, it makes sense for now for the BrokerApiWrapper to remain in multi-site, and this can be revisited in the future if the requirements change.

[1] https://drive.google.com/file/d/1BN4NUkwQ9Y0U00Ox43Px3jzfpyJdeNC1/view?usp=drive_link
[2] https://gerrit-review.googlesource.com/c/modules/events-broker/+/315882
[3] https://gerrit-review.googlesource.com/c/plugins/multi-site/+/304587

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon
Attention: {"person_ident":"Gerrit User 1083454 \u003c1083454@173816e5-2b9a-37c3-8a2e-48639d4f1153\u003e","operation":"REMOVE","reason":"Change was abandoned"}
Attention: {"person_ident":"Gerrit User 1006192 \u003c1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153\u003e","operation":"REMOVE","reason":"Change was abandoned"}
Attention: {"person_ident":"Gerrit User 1072905 \u003c1072905@173816e5-2b9a-37c3-8a2e-48639d4f1153\u003e","operation":"REMOVE","reason":"Change was abandoned"}