GroupCheckerImpl: Mock directly the repository in unit-tests
The unit-tests of GroupCheckerImpl were targeting the internal
behaviour of how the repository works rather than checking directly
the expected outcome. That was working until the JGit internal
behaviour changes and the tests are failing without an apparent
reason.

Mock directly the Repository.exactRef() which is the method
that's expected to be called when the groups are checked.

Change-Id: I156c816803c3d3a87119f24e6839626a07034176