Do not take draft comments into account for change index up-to-date
The change index does not include draft comments.
Also, draft comments are not stored with the change notes
and are user-specific, therefore do not make sense in the
overall computation of the change latest update timestamp.

Previously, the status of the change latest ts was calculated
with the assumption that adding a draft comment would result
in the change status to be modified and therefore needed
to be reindexed up to the latest timestamp.

In multi-site the draft comments are not even replicated
because of the lack of support for draft comments stream
events, added only from Gerrit v3.9.

The consequence was the creation and indefinite retry
of the indexing events, which would have eventally clogged
the logs and the executors.

TODO: This check needs to be reintroduced after the merging to
stable-3.9, with the additional check of the draft comment stream
events being enabled or not.

Bug: Issue 40014838
Change-Id: I2f7e63bd3a97c4463fb4f5417770cbb429026c2b