Update the refs/multi-site/version/value once per batch-refupdate
The update of refs/multi-site/version/value on the global-refdb
is a very expensive operation for two reasons:
- The global-refdb is a remote system, potentially with a high
  network latency
- It needs to maintain global consistency across geographical areas
  which would also include sever-side synchronisation

Reducing the number of updates on the global-refdb is paramount
for keeping an acceptable performance in Gerrit updates.
When multiple updates happen in the same batch for Gerrit, it would
not make sense to update the global-refdb mutliple times for the
version value because globally the batch represents a unit of work
and performing it as a whole means updating the version of the
repository once.

Move from individual stream-events to batch ref-update listener
for updating the refs/multi-site/version/value, so that multiple
updates would result in a single global-refdb update.

For example, the upload of a new patch-set for a change was
previously generating four different updates on the global-refdb
for the refs/multi-site/version/value, whilst now there is only
one single update generated for the  whole batch.

Change-Id: Ia6c7f5aa085a8e2770b40df34145bb937cf1772e