Fix bad merge Ib6aa8a which caused duplicates updates of refs/multi-site/version/*
The merge of Ib6aa8a caused the double-binding of the ProjectVersionRefUpdateImpl
to the stream event listener. This combined with the lack of a Singleton
annotation caused the triggering of a project version update twice per ref-update.

The issue was not present in stable-3.4, because the binding was happening just
once and the lack of the singleton did not have consequences.

Change-Id: Iee3e74bd79bd49cc9dafe9317734def8c8655ef7