Removed repeated comments and silenced wget logs
Since all plugins were echoing on call, moved it to inside the calling
function. Additionally, since the wget when errors echoes a message, it
was silenced. There is a situation of a false error when the first
wget fails.
Also went explicit on the function call to say when we are getting a
plugin

Change-Id: I6ee3c47d7ab723d2c0286163206803934c395699