Do not propagate internal indexing operations
This change brings a feature that was previously integrated into
the High Availability (HA) plugin as part of the change 245113.

Give the ability to block the propagation of the forced reindexing
events (accounts and changes) when there isn't a current request context
available.

The new behaviour is controlled by the index.synchronizeForced
configuration setting.

Bug: Issue 316786570
Change-Id: I1391d268d90b3bf1d4d43b6a9deb4ae64bb6bcda