Take draft comments for change up-to-date when enableDraftCommentEvents is true
Until change I02e17592b78c6b3bf80 draft comments were not replicated across
nodes. This led to an issue with the calculation of the change index
being up-to-date. The problem was addressed and resolved in change
I2f7e63bd3a97c446. Consequently, if
event.stream-events.enableDraftCommentEvents is configured as true, the
code should now utilize draft comments for calculating the change
index's up-to-date status.

Bug: Issue 317084693
Change-Id: I58d0a2d79da90a8cb957e7b517b958f31c4e8c22