Avoid duplicate indexing tasks for the same id
Indexing tasks check if all necessary data is present on the node
by comparing the event timestamp and the event target sha1 with the
local repository sha1. For very active repositories, some of the
indexing tasks never pass this check because:

* During the retry backoff, the target branch of the change was
updated to a newer sha1
* The consumed event is pointing to an outdated `/meta`
NoteDb version.

This means that we can end up with multiple indexing tasks trying to
reindex the same change, some of them except the last one will fail
after the maximum number of retries.

To avoid this situation, make sure that:

* There is only one indexing task trying to index a change.
* If any indexing task successfully indexed a change, previous
indexing task for that change can be discarded.

Bug: Issue 320542020
Change-Id: I0117676ed015209a6b39f05e296adf6caf8b4485