Optimize indexing tasks by squashing duplicates for the same id
Indexing tasks check if all necessary data is present on the node
by comparing the event timestamp and the event target sha1 with the
local repository sha1. For very active repositories, some of the
indexing tasks never pass this check because:

* During the retry backoff, the repository was updated to a newer sha1
* The consumed event is pointing to an already outdated sha1

This means that we can end up with multiple indexing tasks trying to
reindex the same change, and all of them will fail after the maximum
number of retries.

To avoid this situation, make sure that:

* There is only one indexing task retrying to index a change; all
previous tasks can be discarded.
* If any indexing task successfully indexed a change, all previous
indexing tasks for that change can be discarded

Bug: Issue 320542020
Change-Id: Ic1481f2940eda2759c5a8dbd7ac075878050d881