Ignore refs/multi-site/version on global-refdb in push replication filter
When I509dd7075b0 stopped updating the SHA1 for refs/multi-site/version
in global-refdb in this stable-3.4 branch, it did not take into
consideration that existing setups may have plenty of
refs/multi-site/version already stored on them.

Stopping the update of those refs on the global-refdb had a negative
impact on the consistency of the replication of refs/multi-site/version
across the multi-site clusters, causing missed replication events.

Consider the refs/multi-site/version on global-refdb as non-existent
and skip altogether the associated checks if the associated local refs
were already up-to-date.
This change effectively ignores weather the local refs/multi-site/version
is up-to-date or not and always pushes the refs to replication remotes.

This is the completion of the similar fix performed on the
MultisiteReplicationFetchFilter in I8b938675.

Change-Id: I833138849807475a7aed308d8703bb7e16728d2d