Do not filter out refs removals when filtering pull-replication fetches
Since the support of the fetching of refs removals with I3b43c0c,
the async fetch can manage the deletion of refs as part of the replication
task.

The previous assumption that async tasks should not deal with refs
deletions when the global-refdb has a deleted ref isn't valid anymore.

Change-Id: I368e64c47d2cf3b4a74c7054ccb1f06143204d93