Do not throw an exception when refs/multi-site/version lock fails
Busy repositories may have a lot of concurrency on the updates of
refs/multi-site/version: the failure to run the update is not a critical
issue and should not create and log any exception.

Updating the refs/multi-site/version multiple times with the same timestamp
won't make any sense anyway, so the lock failure is expected.

Change-Id: Ib48d0596444103bc11be9233306312f4b27a7067