Disable replication lag tracking by configuration
Introduce the new configuration settings
ref-database.replicationLagEnabled for disabling the metric to trace
the auto-replication lag between sites.

The rationale for disabling this metric lies in the use of
mono-repos with very frequent updates which would cause the ref
refs/multi-site/version/* to be continuously updated causing
packed-refs locks and consuming precious resources on the
global-refdb.

When using pull-replication and apply-object, the value of using the
refs/multi-site/version/* for tracking is not very useful anyway
because it gets replicated outside the batch.

Bug: Issue 313856493
Change-Id: I6b32e0423ed894793f55ed57c60cc356ebe2a31a