Allow to enable/disable push/pull replication filters
The multi-site plugin had always enabled all replication filtering
with the checking of their global-refdb status. Although it is a good
default to have, the lack of flexibility brought some additional
performance issues when running unneeded checks of hundreds of
thousands of refs against the global-refdb.

On Cloud environments where the global-refdb is provided as a service,
the consumption of the APIs is very expensive and can also be capped
or rate-limited, causing potential outages due to unneeded calls.

Introduce two new settings for enabling/disabling the filtering
individually for the push or pull-replication plugins.

Example:

[ref-database]
  pushReplicationFilterClassEnabled = false
  pullReplicationFilterClassEnabled = true

Change-Id: I1989f3bbc8c257f9bb5847e7d3c7755464e2cf3a