Implement filterAndLock in ReplicationFetchFilter
The ReplicationFetchFilter allows to also lock refs whilst filtering
them, which is paramount for preventing the refs to be fetch from being
mutated and therefore causing potential split-brains.

Lock local refs locally for allowing concurrent access to happen on
other nodes that are not impacted by the local replication tasks.

Change-Id: I0f8778dee71b3e3963403541b212058cdd42c784