Reset *ReplicationFilterEnabled automatically at startup
When the push/pull-replication plugin is not available, the filtering
of refs should also be automatically disabled. That was happening
before this change but the associated property was not reset in
the in-memory configuration object, leading other parts of the codebase
assume that the filter was enabled whilst it wasn't.

Change-Id: I5a19d23f75d0b9e2cb4c19705bd2a043f73ff25d