Reset *ReplicationFilterEnabled automatically at startup
When the push/pull-replication plugin is not available, the filtering of
refs should also be automatically disabled. That was happening before
this change, but the associated property was not reset in the in-memory
configuration object, leading other parts of the codebase assume that
the filter was enabled when it wasn't.

For example, when the multi-site.config looks like

[ref-database]
 pullReplicationFilterClassEnabled = true

but there is no pull-replication.jar in the $GERRIT_SITE/plugins then
the pullReplicationFilterEnabled() was returning true whilst that wasn't
the case, as the pull-replication plugin is not loaded and therefore the
filter isn't active and enabled either.

Change-Id: I5a19d23f75d0b9e2cb4c19705bd2a043f73ff25d