Do not fail reindexing changes when its patch-sets aren't replicated yet
Depending on how replication is configured, the changes' /meta refs could
be replicated before its associated patch-sets. Every time that a change
is replicated, the pull-replication plugin emits a fetch-ref-replicated
even associated to the change's refs.

When creating a new change, two refs are replicated:
1. refs/changes/<NN>/<NNNN>/1 (first patch-set of the change)
2. refs/changes/<NN>/<NNNN>/meta (change meta-ref)

When the above refs are replicated out of order, with 2. replicated before
1., then the multi-site plugin tried to reindex the change in an inconsistent
state, throwing a "missing unknown" error and resulting with an stale index.

Instead of reacting only on 'meta' refs replication, consider all
the refs replications events associated with a change and ignore the indexing
errors associated with a missing unknown SHA1.

Bug: Issue 425059627
Change-Id: I9a6337c1cecb52df832d64b2c7bb30e97ba7916a