Use DynamicMap<> for registering routing events handler
Refactoring for reducing the coupling of the event router
receiving events from the events-broker or Gerrit event-bus
to the associated registered handler.

The overall logic and configuration remains unchanged, however,
this change helps decoupling the routing mechanism from the
associated implementation logic, opening the door to a conditional
registration of the events handlers based on configuration.

Change-Id: I194e5571923400ec2755365e5ac9d233ae688dd9