Make index synchronization configurable per index type
Allow configuring which index types should be synchronized across sites
in a multi-primary setup. Supported types are: change, account, group,
and project.

The `index.synchronize` setting accepts a list of types:
  - `change-index`
  - `account-index`
  - `group-index`
  - `project-index`

To disable synchronization entirely set to `false` and
to enable all set to `true`.

This enables finer-grained control on pull-replication and
multi-site indexing mechanism.

For example it will be possible to disable redundant change
indexing retries on fetch-ref-replicated when using pull-replication
in a multi-primary Gerrit setup.

Bug: Issue 425113249
Change-Id: I15fbf51e9814364cf624f251544164ec340ac225